fixes docker run command to report actual result
When we changed the docker run command to remove the container on exit with the
`--rm` flag, we inadvertently stopped the script from fetching the container
exit code and we were always reporting success.

This change reverts the docker run to keep the container and we remove the
container explicitly after fetching the result.