Disable AVNIR and UAVSAR test case 10
We noticed some strange behavior with creating png output
when the input has multiple bands. Any of these requests will return same
output: ` ['Band3', 'Band2', 'Band1'], ['Band2', 'Band1'], ['Band1']` and we
think that the code is only writing the last band in the request to the png.
We will follow up and see if this is functionality that people need or not and
either prevent these types of requests or fix the output in the future.

DAS-1480